move CString implementation to common area
