+++ dlfcn.go
@@ -9,35 +9,23 @@ package purego
 import (
 	"runtime"
 	"unsafe"
+
+	"github.com/ebitengine/purego/internal/strings"
 )
 
 const RTLD_GLOBAL = 0x8
 
 const RTLD_DEFAULT = ^uintptr(1)
 
-// hasSuffix tests whether the string s ends with suffix.
-func hasSuffix(s, suffix string) bool {
-	return len(s) >= len(suffix) && s[len(s)-len(suffix):] == suffix
-}
-
-func cString(name string) *byte {
-	if hasSuffix(name, "\x00") {
-		return &(*(*[]byte)(unsafe.Pointer(&name)))[0]
-	}
-	var b = make([]byte, len(name)+1)
-	copy(b, name)
-	return &b[0]
-}
-
 func Dlopen(name string, mode int) uintptr {
-	bs := cString(name)
+	bs := strings.CString(name, false)
 	ret, _, _ := SyscallN(dlopenABI0, uintptr(unsafe.Pointer(bs)), uintptr(mode), 0)
 	runtime.KeepAlive(bs)
 	return ret
 }
 
 func Dlsym(handle uintptr, name string) uintptr {
-	bs := cString(name)
+	bs := strings.CString(name, false)
 	ret, _, _ := SyscallN(dlsymABI0, handle, uintptr(unsafe.Pointer(bs)), 0)
 	runtime.KeepAlive(bs)
 	return ret
